backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
passlib[argon2]==1.7.4
pyjwt==2.9.0
boto3==1.35.14
redis==5.0.8
pydantic==2.9.1
python-multipart==0.0.9
email-validator==2.2.0
pytest==8.3.4
httpx==0.27.2
